Update N-1 plotting scripts again, add test/NMinus1Effs/nm1entry.py, and move some N-1 specific functions from python/roottools.py to test/NMinus1Effs/nm1entry.py
